seo-article-writer/requirements.txt
requests
beautifulsoup4
openai
python-wordpress-xmlrpc
markdown